+++ apps/nextjs-website/react-components/components/Header/helpers/Header.MenuDropdown.helpers.tsx
@@ -18,7 +18,6 @@ import {
   ExternalLinkIcon,
   isValidExternalLink,
 } from '@react-components/components/common/Common';
-import { usePathname } from 'next/navigation';
 import { useRef } from 'react';
 
 const useStyles = ({ active }: MenuDropdownProp, { spacing }: Theme) => {
@@ -30,7 +29,7 @@ const useStyles = ({ active }: MenuDropdownProp, { spacing }: Theme) => {
       alignItems: 'center',
       margin: 0,
       padding: '0 20px',
-      borderBottom: '3px solid transparent',
+      borderBottom: `3px solid ${active ? '#ffffff' : 'transparent'}`,
     },
     menuItem: {
       height: '100%',
@@ -86,14 +85,6 @@ export const MenuDropdown = (
   const muiTheme = useTheme();
   const styles = useStyles(props, muiTheme);
   const hasLinks = items?.length;
-  const pathname = usePathname();
-  const isCurrentLink = (url?: string) => {
-    if (url && url.indexOf('/') >= 0) {
-      const urlPathname = url.substring(url.indexOf('/'));
-      return pathname === urlPathname;
-    }
-    return false;
-  };
 
   const itemsRef = useRef<Array<HTMLAnchorElement | null>>([]);
 
@@ -176,15 +167,7 @@ export const MenuDropdown = (
   };
 
   return (
-    <Stack
-      sx={{
-        ...styles.menu,
-        ...(isCurrentLink(button.href) && {
-          borderBottomColor: '#ffffff',
-        }),
-      }}
-      component='li'
-    >
+    <Stack sx={styles.menu} component='li'>
       <Box sx={styles.menuItem}>
         <MenuItem />
       </Box>
@@ -203,9 +186,6 @@ export const MenuDropdown = (
                     underline='none'
                     sx={{
                       color: '#5C6F82',
-                      ...(isCurrentLink(item.href) && {
-                        textDecoration: 'underline',
-                      }),
                       fontSize: '1em',
                       fontWeight: 600,
                       padding: 0,
