+++ apps/nextjs-website/react-components/components/Header/helpers/Header.MobileNav.helpers.tsx
@@ -22,7 +22,6 @@ import {
 } from '@react-components/components/common/Common';
 import chevronBlueIcon from '@react-components/assets/icons/icon-chevron-blue.svg';
 import Image from 'next/image';
-import { usePathname } from 'next/navigation';
 
 const drawerStyles = {
   '& .MuiDrawer-paperAnchorLeft': {
@@ -54,15 +53,6 @@ export default function MobileNav({
 
   const toggleMenu = (index: number) => {
     setOpenMenuIndex(openMenuIndex === index ? null : index);
-  };
-
-  const pathname = usePathname();
-  const isCurrentLink = (url?: string) => {
-    if (url && url.indexOf('/') >= 0) {
-      const urlPathname = url.substring(url.indexOf('/'));
-      return pathname === urlPathname;
-    }
-    return false;
   };
 
   const MenuItem = ({
@@ -188,11 +178,7 @@ export default function MobileNav({
                         fontWeight: 600,
                         px: 2,
                         color: '#0066CC',
-                        borderLeftWidth: 3,
-                        borderLeftStyle: 'solid',
-                        borderLeftColor: isCurrentLink(item.href)
-                          ? '#0066cc'
-                          : 'transparent',
+                        borderLeft: `3px solid ${item.active ? '#0066cc' : 'transparent'}`,
                       }}
                     >
                       <MenuItem
